made rows cshow cursor on hover if clickable, fixed bug on propagation, updated short cut buttons so that they remove the row if clicked
